refactor: housecleaning 7 of inf
1. Much of lgrit_math should be considered deprecated with the GLM-binding
library, e.g., ``debug.setmetatable(vector3(0,0,0), glm)``. Which will override
the metamethods for vectors & quaternions with a far more expressive API.

The original lgrit_math source began as a simplified version of the original
grit codebase extended to be feature-complete with lmathlib (which required a
few hundred lines of macro hacking).

... In the meantime, this commit *drastically* optimizes luaVec_trybinTM.

2. vectable_getstr has been rewrote; the previous solution was a
quick-and-dirty hack.