refactor: improve str_blob
Introduce a LUA_TSTRING variant that is effectively a LNGSTR but stored in
tables by reference and not by hash; this handles a specific edge-case not
considered by the previous 'blob' implementation.

The intent is to use the already existing in facilities, e.g., string.pack,
string.unpack, and table.concat, to (safely or safely-as-possible) emulate FFI
structures when blob data is passed & manipulated by external APIs. Allowing
byte data to still be beholden to the garbage collector while not requiring the
allocation of intermediate data/userdata when going to and from the Lua API.

At some point an API similar to Dataview's (functional interfaces of
string.pack/string.unpack will be implemented).