tweak: rework exceptions and API linkage for cpp builds
1. Early refactoring for how Lua manages C++, or links against objects compiled
with C++ (... while Lua is still compiled with C).

2. The current interpreter only uses exceptions and try/catch blocks for
generalized RAII (i.e., ensuring stack-local object destructors are called).

This commit introduces type-specific exception handlers for unprotected calls.
Ensuring the state of the Lua stack is properly maintained throughout the
lifetime of the exception. Ideally, all other exceptions should be considered
as programming errors that lead to panics.

It may be beneficial to include an exception class in luaconf.h that can
be inherited so libraries have the option for Lua to catch exceptions
for them (too much hand-holding).

3. TODO: Introduce a utility class (LuaCrtAllocator) that bridges lua_Alloc and
std::allocator