scripting examples
feat: dataview implementation based on string.blob

feat: example functions tailored to GTA V's script natives

feat: introduce kdtree/octree spatial indexing structures

The spatial indexing structures are examples and not considered to be production
ready/worthy scripts. They have yet to be thoroughly tested or analyzed with a
profiler.

Most query operations require only a couple of microseconds for large datasets.
For a single threaded and interpreted implementation, and without relying on
AOT/JIT compilation, the implementation is as efficient as I hoped it could be.