refactor: vector/matrix C API
Instead of wrapping changes in 'LUA_GRIT_COMPAT', this commit changes the
semantics of invalid vector/matrix indexing to be more aligned with tables.
Maintaining compatibility here is not worth it.

The previous/grit implementations would throw a runtime error on invalid access
(e.g, array-index out of bounds, non-valid index type, etc). Instead, nil is
returned and this interfaces better with the C API.

The README should also be updated with some of the unstated goals of this
runtime. ...