chore(glm/binding): version
Something that resembles semantic versioning but closer to random value
increments in a header.

Mostly grunt tasks remain on the TODO list.

One additional experiment/idea would be to allow the binding library to be
compiled with ALIGNED_GENTYPES while the runtime is not. This will require some
change in headers to properly convey glm::qualifier information.